changed storybook structure for multiple variants
